Rename model to starwars_model
