+++ src/graphql/mod.rs
@@ -1,5 +1,3 @@
-mod model;
+pub mod starwars_model;
 pub mod ocieguide_schema;
 pub mod starwars_schema;
-
-pub use model::*;
