+++ src/graphql/starwars_model.rs
